telegram_contract_bot/requirements.txt
aiogram==3.4.1
python-docx==1.1.0